refactor: decouple jest and story dependencies
