+++ packages/ibm-products/src/components/TagOverflow/TagOverflow.test.js
@@ -12,7 +12,6 @@ import { pkg } from '../../settings';
 import uuidv4 from '../../global/js/utils/uuidv4';
 
 import { TagOverflow } from '.';
-import { FiveTags } from './TagOverflow.stories';
 
 const blockClass = `${pkg.prefix}--tag-overflow`;
 const componentName = TagOverflow.displayName;
@@ -21,6 +20,29 @@ const componentName = TagOverflow.displayName;
 const className = `class-${uuidv4()}`;
 const dataTestId = uuidv4();
 const tagWidth = 60;
+
+const tagLabel = (index) => `Tag ${index + 1}`;
+const tags = Array.from({ length: 20 }, (v, k) => ({
+  label: tagLabel(k),
+  id: `id-${k}`,
+}));
+
+const fiveTags = tags.slice(0, 5);
+
+const tagOverflowProps = {
+  items: fiveTags,
+};
+
+const FiveTags = (argsIn) => {
+  const { containerWidth, ...args } = {
+    ...argsIn,
+  };
+  return (
+    <div style={{ width: containerWidth }}>
+      <TagOverflow {...args} />
+    </div>
+  );
+};
 
 describe(componentName, () => {
   const { ResizeObserver } = window;
@@ -80,13 +102,14 @@ describe(componentName, () => {
   });
 
   it('Renders all as visible tags when space available', async () => {
-    const tagCount = FiveTags.args.items.length;
+    const tagCount = tagOverflowProps.items.length;
+
     window.innerWidth = tagWidth * tagCount + 40;
 
-    render(<FiveTags {...FiveTags.args} />);
+    render(<FiveTags {...tagOverflowProps} />);
 
-    const firstTagLabel = FiveTags.args.items[0].label;
-    const lastTagLabel = FiveTags.args.items[tagCount - 1].label;
+    const firstTagLabel = tagOverflowProps.items[0].label;
+    const lastTagLabel = tagOverflowProps.items[tagCount - 1].label;
 
     // first and last should be visible
     screen.getByText(firstTagLabel, {
@@ -99,7 +122,7 @@ describe(componentName, () => {
   });
 
   it('Obeys max visible', async () => {
-    render(<FiveTags {...FiveTags.args} maxVisible={3} />);
+    render(<FiveTags {...tagOverflowProps} maxVisible={3} />);
 
     expect(
       screen.getAllByText(/Tag [0-9]+/, {
@@ -111,7 +134,7 @@ describe(componentName, () => {
   // The below test case is failing due to ResizeObserver mock
   // it('Renders only the overflow when very little space', async () => {
   //   window.innerWidth = tagWidth / 2;
-  //   render(<FiveTags {...FiveTags.args} />);
+  //   render(<FiveTags {...tagOverflowProps} />);
 
   //   const visible = screen.queryAllByText(/Tag [1-5]+/, {
   //     selector: `.${blockClass}__item--tag span`,
